FindBugs code cleanups (4/13) - CoreConnection.
CoreConnection: Make the class final so we don't have to worry about subclasses having the read thread start before their constructor is even called (the alternative would be to introduce a new method to start the read thread and pull that out of the constructor.)

Also, replace uses of Map.keySet() that proceed to use map.get(key), because this adds another O(n*log(n)) lookup. The entrySet *gives* you the value in O(1) time, and the entrySet is actively maintained when the map is mutated, so returning it to the caller is also an O(1) operation, so why not use it?

Also, replace instances of repeated string concatenation in a loop with a StringBuilder. This cuts down on object construction costs (memory allocations), while flattening the quadratic cost of repeated string concatenation down to linear time. StringBuilder is not thread-safe, so it actually has less overhead than StringBuffer (we don't need thread safety in this context).